requirement.txt
backports-abc==0.5
futures==3.2.0
pymongo==3.6.1
python-slugify==1.2.5
singledispatch==3.4.0.3
six==1.11.0
tornado==5.0.2
Unidecode==1.0.22
uuid==1.30
